SOAP/requirements.txt
spyne == 2.14.0
suds-jurko == 0.6
lxml == 4.9.2
BeautifulTable == 1.1.0